Relocate framework support files + port script + timeout fallbacks (Phase 3, batch 13)
Issue #683 (parent #680, Phase 3) batch 13: migrate the support /
port script materializers plus the two timeout-fallback no-op
predicates from `turn.rs` to `scaffold_pipeline.rs` via the veneer
pattern.

Migrated as free fns in `scaffold_pipeline.rs` (`pub(super)`):

* `maybe_apply_deterministic_framework_support_files(&Agent, request)
    -> Result<(), String>`
* `maybe_apply_requested_port_script(&Agent, request)
    -> Result<(), String>`
* `maybe_apply_deterministic_quality_fallback_after_timeout(
    &Agent, err) -> Option<AssistantReply>`
* `maybe_apply_deterministic_polish_fallback_after_timeout(
    &Agent, err) -> Option<AssistantReply>`

Helper kept module-private:

* `write_react_dev_wrapper(&Agent, wrapper)` — was a private `fn` on
  `impl Agent`; moved as a private free fn in `scaffold_pipeline.rs`
  (single caller `maybe_apply_requested_port_script` migrated
  alongside it).

Promoted in `turn.rs` to `pub(super)`:

* `sync_package_json_with_existing_lock`

`scaffold_pipeline.rs` imports gain
`super::quality::{package_json_with_requested_port,
react_dev_wrapper_for_requested_port}` and
`super::turn::sync_package_json_with_existing_lock`. The interior
caller `maybe_apply_deterministic_quality_fallback` now invokes
sibling free fns directly instead of going through `agent.<method>()`.

Removed (dead veneer, sole caller migrated in this PR):

* `Agent::maybe_apply_deterministic_framework_support_files`

`turn.rs` adjustments:

* 3 migrated method slots stay as one-line veneer shims so caller
  sites (`run_actor_loop` recovery + timeout paths) stay unchanged.
* Pruned now-unused `package_json_with_requested_port` /
  `react_dev_wrapper_for_requested_port` from the top-of-file
  `super::quality::{...}` import block and
  `deterministic_framework_game_impl_path` /
  `deterministic_support_target_relative` from the top-of-file
  `super::scaffold_pipeline::{...}` import block.

LOC delta:

* `turn.rs`: 30,543 → 30,438 (-105 LOC)
* `scaffold_pipeline.rs`: 1,508 → 1,612 (+104 LOC)
* Cumulative `turn.rs`: 39,489 → 30,438 (-9,051 LOC, -22.9%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
